Creates ConnectivityManager extension component
