+++ app/src/main/java/org/mozilla/fenix/share/ShareViewModel.kt
@@ -23,7 +23,7 @@ import kotlinx.coroutines.launch
 import mozilla.components.concept.sync.DeviceCapability
 import mozilla.components.service.fxa.manager.FxaAccountManager
 import org.mozilla.fenix.ext.components
-import org.mozilla.fenix.settings.isOnline
+import org.mozilla.fenix.ext.isOnline
 import org.mozilla.fenix.share.listadapters.AppShareOption
 import org.mozilla.fenix.share.listadapters.SyncShareOption
