+++ app/src/test/java/org/mozilla/fenix/share/ShareViewModelTest.kt
@@ -26,7 +26,7 @@ import org.junit.runner.RunWith
 import org.mozilla.fenix.TestApplication
 import org.mozilla.fenix.ext.application
 import org.mozilla.fenix.ext.components
-import org.mozilla.fenix.settings.isOnline
+import org.mozilla.fenix.ext.isOnline
 import org.mozilla.fenix.share.listadapters.AppShareOption
 import org.mozilla.fenix.share.listadapters.SyncShareOption
 import org.robolectric.RobolectricTestRunner
